+++ src/app/js/events.js
@@ -1,3 +1,5 @@
 const EventEmitter = require("events");
 class Emitter extends EventEmitter {};
 const events = new Emitter();
+
+module.exports = events;
