+++ src/app/js/set_dom_strings.js
@@ -1,12 +1,12 @@
-// Replaces strings in the HTML will language strings properly. This
+// replaces strings in the HTML will language strings properly. This
 // searches for %%<string>%%, aka, %%gui.exit%% will be replaced with
 // "Exit", this works without issues.
-function setlang() {
-	// Finds %%%% strings
+module.exports = () => {
+	// finds %%%% strings
 	html = document.body.innerHTML.split("%%");
 
 	for (let i = 0; i < html.length; i++) {
-		// Simply checks to make sure it is actually a lang string.
+		// simply checks to make sure it is actually a lang string.
 		if (html[i][0] != " " && 
 			html[i][html[i].length - 1] != " ") {
 			// Replaces it with it's string
@@ -14,6 +14,6 @@ function setlang() {
 		}
 	}
 	
-	// Replaces the original HTML with the translated/replaced HTML
+	// replaces the original HTML with the translated/replaced HTML
 	document.body.innerHTML = html.join("");
 }
